refactor: change some methods to protect the coding convention
